requirements.txt
torch
torchaudio
numpy
sounddevice
python-dotenv
deepgram-sdk
groq
edge-tts
scipy
soundfile
packaging